Refactor List/ListDir
WIP: DO NOT MERGE. Tests are expected to fail.

Concept proposal for #316

This shows the concept. All the heavy lifting is done in the "fs" package, making file system implementations a bit simpler.

/fs shows interfaces, and should compile.
/amazonclouddrive: Changed to new interface.
/b2: Changed to new interface.
/local: Changed to new interface.

The Fs no longer needs to create a separate goroutine when List/ListDir is called. Instead it can block as long as it likes. This simplifies implementations a bit (at least it doesn't become more complex).

Implementation is hidden in "fs" package, and only exposes the minimally needed interface, which of course can be extended.